requirements.txt
numpy==1.17.4
adjustText==0.7.3
alembic==1.3.1
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
blinker==1.4
bokeh==1.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configobj==5.0.6
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
defusedxml==0.6.0
develop==0.1.4
docutils==0.16
email-verifier==1.0.0
entrypoints==0.3
et-xmlfile==1.0.1
fastcluster==1.1.26
Flask==1.1.1
Flask-EmailVerifier==0.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.2
fuzzywuzzy==0.18.0
gunicorn==20.0.4
humanize==0.5.1
idna==2.8
importlib-metadata==1.4.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.1.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
matplotlib-venn==0.11.5
mistune==0.8.4
mpld3==0.3
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
openpyxl==3.0.3
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.0
pathlib2==2.3.5
pexpect==4.8.0
phantomjs==1.1.4
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psutil==5.7.0
ptyprocess==0.6.0
pycairo==1.19.0
pycrypto==2.6.1
Pygments==2.5.2
PyGObject==3.30.4
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.5
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-hglib==2.6.1
python-Levenshtein==0.12.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
readme-renderer==24.0
redis==3.3.11
repo==0.1.0
requests==2.22.0
retrying==1.3.3
ruamel.appconfig==0.5.4
ruamel.showoutput==0.1.7
ruamel.std.argparse==0.8.1
scikit-learn==0.22
scipy==1.3.3
seaborn==0.10.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
setuptools==40.8.0
sklearn==0.0
SQLAlchemy==1.3.11
suds-jurko==0.6
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
waitress==1.4.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
wheel==0.32.3
widgetsnbextension==3.5.1
WTForms==2.2.1
xlrd==1.2.0
zipp==2.1.0
